Add files via upload
This is my first repository on github. Today, I started a data science course on Coursera and I was recommended that i should upload my first jupyter notebook to the Github and i did it. I am extremely delighted that I learnt a new thing today, I am liking this platform and I will be exploring more of it from now on.